requirements.txt
tensorflow==2.12.0
numpy==1.23.5
Pillow==9.5.0
opencv-python-headless==4.7.0.72
Flask==2.2.5
gunicorn

